+++ core/object/node_keywords.go
@@ -532,11 +532,9 @@ var nodeCommonKeywords = []keywords.Keyword{
 		Text:        keywords.NewText(fs, "text/kw/node/cluster.id"),
 	},
 	{
-		Section:  "cluster",
-		Option:   "name",
-		Scopable: true,
-		Default:  "default",
-		Text:     keywords.NewText(fs, "text/kw/node/cluster.name"),
+		Section: "cluster",
+		Option:  "name",
+		Text:    keywords.NewText(fs, "text/kw/node/cluster.name"),
 	},
 	{
 		Section:     "cluster",